refactor: component 이름 수정 및 styled 파일 분리 (#10) 	deleted:    src/components/Layout.tsx 	deleted:    src/components/Table.tsx 	renamed:    src/components/common/btn/cancelBtn/CancelBtn.component.tsx -> src/components/common/btn/cancelBtn/CancelBtn.tsx 	modified:   src/components/index.ts 	new file:   src/components/layout/Layout.styled.ts 	new file:   src/components/layout/Layout.tsx 	new file:   src/components/table/Table.styled.ts 	new file:   src/components/table/Table.tsx 	renamed:    src/pages/article/create/ArticleCreate.component.tsx -> src/pages/article/create/ArticleCreate.tsx 	renamed:    src/pages/article/detail/ArticleDetail.component.tsx -> src/pages/article/detail/ArticleDetail.tsx 	renamed:    src/pages/article/main/Article.component.tsx -> src/pages/article/main/Article.tsx 	deleted:    src/pages/article/main/container/ArticleTable.tsx 	new file:   src/pages/article/main/container/ArticleTable/ArticleTable.styled.ts 	new file:   src/pages/article/main/container/ArticleTable/ArticleTable.tsx 	renamed:    src/pages/article/update/ArticleUpdate.component.tsx -> src/pages/article/update/ArticleUpdate.tsx 	modified:   src/pages/index.ts
