+++ src/components/layout/Layout.styled.ts
@@ -1,0 +1,22 @@
+import styled from "styled-components";
+
+export const Root = styled.main`
+  display: flex;
+  justify-content: start;
+  flex-direction: column;
+  align-items: center;
+  width: 100%;
+  height: 100vh;
+  padding: 20px 0;
+`;
+
+export const Header = styled.header`
+  width: 80%;
+  display: flex;
+  justify-content: start;
+  margin-bottom: 2rem;
+`;
+
+export const Section = styled.section`
+  width: 80%;
+`;
